할일과 날짜 변경하기

+++ ts-todo-list/src/components/TodoList.tsx
@@ -1,8 +1,9 @@
 import { useDispatch } from 'react-redux';
-import { toggle, remove } from '../modules/todo';
+import { toggle, remove, change } from '../modules/todo';
 import { faCheckCircle as checkedCricle } from '@fortawesome/free-solid-svg-icons'
 import { faCheckCircle } from '@fortawesome/free-regular-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { useState } from 'react';
 
 
 interface Props {
@@ -15,21 +16,70 @@ interface Props {
 }
 
 export const TodoList: React.FC<Props> = ({ data }) => {
+    const [titleChange, setTitleChange] = useState({ TF: true, value: data.title });
+    const [dateChange, setDateChange] = useState(true);
+
     const dispatch = useDispatch();
+
     const onClick = () => dispatch(toggle(data.id));
+
     const Delete = (e: any) => {
         e.preventDefault();
-        if (window.confirm("정말 삭제하시겠습니까?")) {
-            console.log("삭제", data.id)
-            dispatch(remove(data.id))
+        if (window.confirm("정말 삭제하시겠습니까?")) dispatch(remove(data.id))
+    }
+
+    const Change = (e: React.MouseEvent<HTMLParagraphElement>) => {
+        if (e.currentTarget.id === "title") {
+            setTitleChange({
+                ...titleChange,
+                TF: false
+            });
+        }
+
+        else if (e.currentTarget.id === "date") {
+            setDateChange(false);
+        }
+    }
+
+    const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+        if (e.target.name === "title") {
+            setTitleChange({
+                ...titleChange,
+                value: e.target.value
+            });
+        } else if (e.target.name === "date") {
+            setDateChange(true);
+            dispatch(change(data.id, "", e.target.value));
+        }
+    }
+
+    const onKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
+        if (e.key === 'Enter') {
+            setTitleChange({
+                ...titleChange,
+                TF: true
+            });
+            dispatch(change(data.id, titleChange.value));
         }
     }
 
     return (
         <li onContextMenu={Delete}>
-            <h3 className={data.toggle ? "line" : ""}>{data.title}</h3>
+
+            {titleChange.TF ? (
+                <h3 onDoubleClick={Change} id="title" className={data.toggle ? "line" : ""}>{data.title}</h3>
+            ) : (
+                <input type="text" value={titleChange.value} onKeyPress={onKeyPress} className="changeInput" name="title" onChange={onChange} />
+            )}
+
             <div className="information">
-                <p className={data.toggle ? "line" : ""}>{data.date}</p>
+
+                {dateChange ? (
+                    <p onDoubleClick={Change} id="date" className={data.toggle ? "line" : ""}>{data.date}</p>
+                ) : (
+                    <input type="date" name="date" onChange={onChange} />
+                )}
+
                 {data.toggle ? (
                     <FontAwesomeIcon onClick={onClick} className="check" size='lg' icon={checkedCricle} />
                 ) : (
